Primer commit

+++ Inmobiliaria/Migrations/InmobiliariaContextModelSnapshot.cs
@@ -247,8 +247,8 @@ namespace Inmobiliaria.Migrations
                     b.Property<double>("Cuit")
                         .HasColumnType("float");
 
-                    b.Property<double>("Dni")
-                        .HasColumnType("float");
+                    b.Property<int>("Dni")
+                        .HasColumnType("int");
 
                     b.Property<string>("Domicilio")
                         .IsRequired()
@@ -288,7 +288,7 @@ namespace Inmobiliaria.Migrations
                             CondicionIva = 2,
                             Cuil = 20361962595.0,
                             Cuit = 0.0,
-                            Dni = 36196259.0,
+                            Dni = 36196259,
                             Domicilio = "Juan Peron Y urquiza",
                             Email = "inmobiliariajuliandaniel@gmail.com",
                             EstadoCivil = 1,
